backend/requirements.txt
Django==5.0.14
djangorestframework==3.16.1
django-cors-headers==4.3.1
django-mptt==0.16.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.0.0
dj-database-url==2.1.0
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2